added RPC coomand & CLI command
Added “stop_wallet” to wallet RPC interface to allow the graceful storage and shutdown of a wallet daemon from the json RPC API.
Added “—exit-after-generate” switch to the simplewallet cli because there was no way to generate a new wallet and then exit the application. “—command” does not work with exit because of the way the application is structured.