Derive view key from spend key
Previously, both the private spend and view key were generated
randomly, and were completely unrelated. However, this method
doesn't allow us to generate a mnemonic seed, at least by the
standard monero method. This also means users will no longer
need to save both a spend and view key, and instead will only
need to send a spend key (which is the same as a seed), or a
mnemonic seed.

This commit causes the view key to be generated by hashing the
spend key with keccak-256, then using this as a seed to generate
a new set of keys, the private and public view key.

This is the same method the paper-wallet uses.