Output mnemonic seed for new wallets
As before mentioned, the new wallets use the deterministic
method of generating seeds, and so we can now give them their
seed. However, we'll only do it when generating new wallets and
not give an option so users don't think they can reimport
from a seed with an old wallet.