Allow imports of different language seeds
Haven't added export yet, but this allows us to import a seed
from any language that is supported - currently:

static std::string languages[] = {"English", "Nederlands", "Français", "Português", "Italiano", "Deutsch", "русский язык", "简体中文 (中国)", "Esperanto", "Lojban"};

The current issue is that if we do it this way, we can't print out
words that aren't in the chosen language dictionary - we would
have to ask for a language argument for this. I'm also unsure
if any languages share enough words that it would be possible
to import a different languages seed from the one you generated.